+++ Sources/KeyboardKit/Callouts/Providers/StandardCalloutActionProvider.swift
@@ -10,8 +10,7 @@ import Foundation
 
 /**
  This provider is initialized with a collection of localized
- providers, and will use the one with the same locale as the
- provided ``KeyboardContext``.
+ providers, and will try to use one that matches the context.
  */
 open class StandardCalloutActionProvider: CalloutActionProvider {
 
@@ -20,58 +19,38 @@ open class StandardCalloutActionProvider: CalloutActionProvider {
 
       - Parameters:
         - keyboardContext: The keyboard context to use.
-        - providers: The action providers to use.
+        - providers: The action providers to use, by default `empty`.
      */
     public init(
         keyboardContext: KeyboardContext,
-        localizedProviders: [CalloutActionProvider & LocalizedService] = [standardProvider]
+        localizedProviders: [CalloutActionProvider & LocalizedService] = []
     ) {
         self.keyboardContext = keyboardContext
         let dict = Dictionary(uniqueKeysWithValues: localizedProviders.map { ($0.localeKey, $0) })
         self.localizedProviders = LocaleDictionary(dict)
     }
-    
-    /**
-     Get the standard callout action provider.
 
-     This can be set to change the standard value everywhere.
-     */
-    public static var standardProvider: CalloutActionProvider & LocalizedService {
-        guard let provider = try? EnglishCalloutActionProvider() else { fatalError("EnglishCalloutActionProvider could not be created.") }
-        return provider
-    }
-
-    /**
-     The keyboard context to use.
-     */
+    /// The keyboard context to use.
     public let keyboardContext: KeyboardContext
     
-    /**
-     This is used to resolve the a provider for the context.
-     */
+    /// This is used to resolve the a provider for the context.
     public var localizedProviders: LocaleDictionary<CalloutActionProvider>
 
 
-    /**
-     Get callout actions for the provided `action`.
-     */
+    /// Get callout actions for the provided action.
     open func calloutActions(for action: KeyboardAction) -> [KeyboardAction] {
         let provider = provider(for: keyboardContext)
         let actions = provider.calloutActions(for: action)
         return actions
     }
 
-    /**
-     Get the provider to use for the provided `context`.
-     */
+    /// Get the provider to use for the provided context.
     open func provider(for context: KeyboardContext) -> CalloutActionProvider {
         provider(for: context.locale)
     }
     
-    /**
-     Get the provider to use for the provided `locale`.
-     */
+    /// Get the provider to use for the provided locale.
     open func provider(for locale: Locale) -> CalloutActionProvider {
-        localizedProviders.value(for: locale) ?? Self.standardProvider
+        localizedProviders.value(for: locale) ?? .disabled
     }
 }