+++ RELEASE_NOTES.md
@@ -20,6 +20,7 @@ These release notes will only contain the current version. Just check out an old
 ### 💥 Breaking changes 
 
 * All deprecated code has been removed or addressed.
+* `StandardCalloutActionProvider.standardProvider` has been removed.